build: enable mypy for tools and fix resulting type errors
Extend [tool.mypy].files across tools, fund, forecast_snapshot, trigger,
and ocr, then resolve the type-check failures that surface:

- paper_trading_cli.py: type the parser error override as Never, cast
  requests JSON responses at each endpoint boundary, rename shadowed
  kwargs, and cast Namespace subcommand access
- proxy_demo.py: type request params as str|int|bool
- grid_trading_screener.py: cast untyped tushare DataFrame results
- stock_return_report.py: convert pandas scalar to float before round()
- stock_monitor_cli.py: validate stock_code before name resolution
- monitor_fallback_stocks.py: use is_a_market_open_today wrapper
- monitor test doubles: replace append-expression mocks with typed
  callbacks and contain generic DataFrame overrides

Update AGENTS.md to document adding modules to the mypy file list.

+++ tools/stock_return_report.py
@@ -85,7 +85,7 @@ def _calculate_return(history_df: pd.DataFrame, lookback: int) -> float | None:
     if pd.isna(latest_close) or pd.isna(base_close) or base_close == 0:
         return None
 
-    return round((latest_close / base_close - 1) * 100, 2)
+    return round(float((latest_close / base_close - 1) * 100), 2)
 
 
 def _format_for_print(report_df: pd.DataFrame) -> pd.DataFrame:
